mem-benchmarks/base/requirements.txt
-r requirements_scenario.txt

wheel
memray
pyperf
pyyaml
attrs
httpretty==0.9.7
